SCSS FIRST

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 import { Console } from "console";
 import Video from "../models/video";
+import User from "../models/user";
 
 // export const videoTrend = (req, res) => {
 //   Video.find({},(error, vids) => {
@@ -7,7 +8,9 @@ import Video from "../models/video";
 //   });
 // }
 export const videoTrend = async (req, res) => {
-  const vids = await Video.find({}).sort({ createdAt: "asc" });
+  const vids = await Video.find({})
+    .sort({ createdAt: "asc" })
+    .populate("owner");
   return res.render("home", { pageTitle: "HOME", vids });
 };
 export const videoSearch = async (req, res) => {
@@ -23,18 +26,23 @@ export const videoSearch = async (req, res) => {
   res.render("search", { pageTitle: "SEARCH VIDEO", videos });
 };
 export const videoGetUpload = (req, res) => {
-  return res.render("upload", { pageTitle: "UPLOAD VIDEO" });
+  res.render("upload", { pageTitle: "UPLOAD VIDEO" });
 };
 export const videoPostUpload = async (req, res) => {
-  const { title, description, hashtags } = req.body;
   const { path: fileUrl } = req.file;
+  const { user: _id } = req.session;
+  const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       fileUrl,
       title,
       description,
+      owner: _id,
       hashtags: Video.makeHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.video.push(newVideo._id);
+    await user.save();
     return res.redirect("/");
   } catch (error) {
     console.log(error);
@@ -46,28 +54,43 @@ export const videoPostUpload = async (req, res) => {
 };
 export const videoSee = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   console.log(video);
   if (!video) {
     res.render("404", { pageTitle: "VIDEO NOT FOUND" });
   } else {
-    res.render("watch", { pageTitle: `WATCHING : ${video.title}`, video });
+    res.render("watch", {
+      pageTitle: `WATCHING : ${video.title}`,
+      video,
+    });
   }
 };
 export const videoGetEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
-    res.render("404", { pageTitle: "VIDEO NOT FOUND" });
+    res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND" });
+  }
+  if (String(_id) !== String(video.owner)) {
+    res.status(403).redirect("/");
   } else {
     res.render("edit", { pageTitle: `EDITING : ${video.title}`, video });
   }
 };
 export const videoPostEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.exists({ _id: id });
   if (!video) {
     res.render("404", { pageTitle: "VIDEO NOT FOUND" });
+  }
+  if (String(_id) !== String(video.owner)) {
+    res.status(403).redirect("/");
   } else {
     const { title, description, hashtags } = req.body;
     await Video.findByIdAndUpdate(id, {
@@ -80,6 +103,17 @@ export const videoPostEdit = async (req, res) => {
 };
 export const videoDelete = async (req, res) => {
   const { id } = req.params;
-  await Video.findByIdAndDelete(id);
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND" });
+  }
+  if (String(_id) !== String(video.owner)) {
+    res.status(403).redirect("/");
+  } else {
+    await Video.findByIdAndDelete(id);
+  }
   return res.redirect("/");
 };
